Get very basic virtual transport working
This implementation takes a few steps back and undoes the
work done to implement an async handler. The work was previously
done in a way that was not natural to celery, so we'll gradually
work back toward that using the blessed approach. The current
implementation should provide pretty complete virtual transport support.
It does not use topics as topics, but creates very bare-bones queues.